Implement createDiscs() and renderDiscs() methods to create the disc objects and render them
The createDiscs() method has been implemented to create the objects representing the discs with width and height based on the number of discs. It will set the "discs" state to have all of the discs at the first index (on the first peg) with the other indices (pegs) empty. The renderDiscs() method has also been implemented to return the Animated.Views of the discs.