Add TouchableOpacities for each of the pegs
Touchable Opacities have been added for each of the pegs. The onPress calls the lift() function or the drop() function depending on if a disc has already been lifted. The blocked state prevents a player from lifting another disc while the disc lifting animation is still running. The lift() and drop() functions have yet to be implemented.