Implement the flashing of the Touchable Opacities to indicate that an illegal move has been attempted
The background color of the Touchable Opacities is now linked to the redFlash state. The flash(index) function will set the redFlash state to be an index of a Touchable Opacity so that the background color of that TO is red for a short period of time to indicate an illegal move.