Add move counter and connect game mode to Settings component via route.params
There is now a move counter the counts the number of moves taken to solve the puzzle. The settings game mode selector is now connected to the game component via route.params. This allows the user to select between "Zen Mode" (just the puzzle), "Count Moves" (the moves taken to solve are counted), and "Timed" (the puzzle solve is timed).