Build tutorial overlay proof-of-concept, delete Tutorial component
The tutorial will be an overlay of the game. The tutorial overlay consists of 4  translucent animated views that cover the entire screen except for a clear box that is specified by its top, left, height, and width. This is used to highlight part of the screen (the negative space not covered by the views) and to only allow the user to interact with that portion of the screen. There are buttons to click through the text instructions, and the buttons will animate the negative space box to highlight the part of the game relating to that instruction in the tutorial.