Implement interactive tutorial with instructions and moving window
The tutorial has now been implemented. There are text instructions that the user can click through by clicking an "okay" button. At some parts in the tutorial, the user continues by interacting with the game. There are animated views that restrict what the user can interact with, and they guide the user to tap the screen in the correct places at the correct times.